- Lazy loading test coverage for editing a resource - Fixed data setter for lazy loading
